requirements.txt
Flask==2.3.2
Flask-Cors==3.2.2
google-cloud-storage==2.8.1
google-cloud-firestore==2.11.0
tensorflow==2.11.0
opencv-python==4.7.0.72
pytesseract==0.3.9
numpy==1.24.2
werkzeug==2.3.4
